refactor(ci): compress queue latency tooling
