+++ scripts/run-core-affected-native.mjs
@@ -89,7 +89,6 @@ const sdkQualificationPaths = [
   'pnpm-lock.yaml',
   'pnpm-workspace.yaml',
   'product/scripts/archive.mjs',
-  'scripts/affected-native-cache-payload.mjs',
   'scripts/affected-native-proof.mjs',
   'scripts/candidate-timeline-events.cjs',
   'scripts/generate-layered-sdk.mjs',
@@ -471,10 +470,9 @@ export function planFromChanged(
     'framework/core/uv.lock',
     'framework/core/package.json',
     'framework/core/tests/',
-    'scripts/affected-native-cache-payload.mjs',
+    'scripts/write-affected-native-cache-manifests.mjs',
     'scripts/run-core-affected-native.mjs',
     'scripts/affected-native-proof.mjs',
-    'scripts/write-affected-native-cache-manifests.mjs',
     '.github/workflows/affected-native-cache-promote.yml',
     '.github/workflows/affected-native-pr.yml',
     'shifu.gates.json',
